pyproject.toml
[project]
name = "device-statistics"
version = "0.1.0"
description = ""
authors = [
    {name = "Max Pfeiffer",email = "max@maxpfeiffer.ch"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.121.1,<0.122.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "alembic (>=1.17.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "semver (>=3.0.4,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "furl (>=2.1.4,<3.0.0)"
]

[dependency-groups]
dev = [
    "pre-commit (>=4.4.0,<5.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest (>=9.0.1,<10.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
src = ["app", "tests"]
extend-exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "E", # pycodestyle
    "W", # pycodestyle
    "I", # isort
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "ASYNC", # flake8-async
    "RUF",  # Ruff-specific rules
]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.pytest]
testpaths = [
    "tests",
]
